refactor: enhance JWT refresh logic and improve email handling; add utility methods for date conversion and streamline user preference management

+++ Backend/src/main/java/com/mdevs/trackera/service/UserEmailService.java
@@ -30,8 +30,9 @@ public class UserEmailService {
         return userEmail;
     }
 
-    public void deleteIfNotLinkedToOAuth(UserEmail userEmail) {
-        if (userOAuthProviderRepository.notExistsByUserAndProviderEmail(userEmail.getUser(), userEmail)) {
+    public void deleteIfUnused(UserEmail userEmail) {
+        User user = userEmail.getUser();
+        if (!user.isEmailLinked(userEmail) && userOAuthProviderRepository.notExistsByUserAndProviderEmail(userEmail.getUser(), userEmail)) {
             userEmailRepository.delete(userEmail);
         }
     }
